+++ CHANGES.rst
@@ -182,7 +182,7 @@ astropy.visualization
 astropy.wcs
 ^^^^^^^^^^^
 
-- Added a ``astropy.wcs.wcsapi.efficient_pixel_to_pixel`` function that can be used
+- Added a ``astropy.wcs.wcsapi.pixel_to_pixel`` function that can be used
   to transform pixel coordinates in one dataset with a WCS to pixel coordinates in
   another dataset with a different WCS. This function is designed to be efficient
   when the input arrays are broadcasted views of smaller arrays. [#9209]